# docs/swagger.yaml
definitions:
  dto.AuthResponse:
    properties:
      hasStore:
        type: boolean
      role:
        type: string
      storeId:
        type: integer
      token:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
  dto.CartItem:
    properties:
      product:
        $ref: '#/definitions/dto.Product'
      quantity:
        type: integer
    type: object
  dto.CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CartItem'
        type: array
      total_price:
        type: number
    type: object
...
  dto.Product:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_boosted:
        type: boolean
      is_posted:
        type: boolean
      name:
        type: string
      price:
        type: number
      seller_id:
        type: integer
      stock:
        type: integer
      store_id:
        type: integer
    type: object
...
  errorx.AppError:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        $ref: '#/definitions/errorx.ErrorCode'
      message:
        type: string
    type: object
  errorx.ErrorCode:
    enum:
    - UNAUTHORIZED
    - FORBIDDEN
    - VALIDATION_ERROR
    - NOT_FOUND
    - INTERNAL_ERROR
    - BAD_REQUEST
    - CONFLICT
    type: string
    x-enum-varnames:
    - ErrUnauthorized
    - ErrForbidden
    - ErrValidation
    - ErrNotFound
    - ErrInternal
    - ErrBadRequest
    - ErrConflict
  response.BaseResponse:
    properties:
      data: {}
      error: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /auth/telegram:
    post:
      consumes:
      - application/json
      description: Validates initData and returns JWT
      parameters:
      - description: initData
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Authenticate via Telegram
      tags:
      - Auth
...
  /user/cart:
    get:
      description: Fetch the active user's cart including product details and total
        price
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Get user cart
      tags:
      - Cart
  /user/cart/add:
    post:
      description: Add a product to the cart with a specific quantity
      parameters:
      - description: Product ID
        in: query
        name: product_id
        required: true
        type: integer
      - description: Quantity
        in: query
        name: quantity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Add to cart
      tags:
      - Cart
  /user/cart/clear:
    delete:
      description: Empty the user's active cart completely
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Clear cart
      tags:
      - Cart
  /user/cart/remove:
    delete:
      description: Remove a specific item from the cart entirely
      parameters:
      - description: Product ID
        in: query
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Remove from cart
      tags:
      - Cart
  /user/cart/update:
    put:
      description: Update the specific quantity of a cart item
      parameters:
      - description: Product ID
        in: query
        name: product_id
        required: true
        type: integer
      - description: New Quantity
        in: query
        name: quantity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                error:
                  $ref: '#/definitions/errorx.AppError'
              type: object
      summary: Update cart item
      tags:
      - Cart
...
swagger: "2.0"